requirements.txt
streamlit==1.25.0
tensorflow==2.13.0
numpy==1.22.0
Pillow==9.0.1
matplotlib==3.5.1

